refactor: 移除 BeatmapDbOperator，数据库操作全部改用 OsuPlayerDbContext
删除 `BeatmapDbOperator` 静态扩展类，其所有方法已迁移至 `OsuPlayerDbContext` 的实例方法。更新所有调用方（ObservablePlayController、SafeDbOperator、OsuFileScanner、IdentifiableExtension、OsuDbInst）直接创建并使用 `OsuPlayerDbContext` 实例执行数据库操作。此举彻底清除遗留的代理层，统一数据库访问方式，无破坏性变更。
