+++ OsuPlayer.Data/TimeSortMode.cs
@@ -1,0 +1,8 @@
+namespace Milky.OsuPlayer.Data
+{
+    public enum TimeSortMode
+    {
+        PlayTime,
+        AddTime
+    }
+}
